requirements.txt
aiohttp==3.12.14
beautifulsoup4==4.13.5
chardet==5.2.0
defusedxml==0.7.1
html2text==2025.4.15
mammoth==1.9.1
markdownify==1.1.0
numpy==2.3.3
omegaconf==2.3.0
onnxruntime==1.22.1
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
pandas==2.3.2
Pillow==11.3.0
pyclipper==1.3.0.post6
pydantic==2.11.9
pymupdf==1.26.4
python_pptx==1.0.2
Requests==2.32.5
Shapely==2.1.1
transformers==4.56.0
